requirements.txt
fastapi==0.110.0
uvicorn==0.27.1

mysql-connector-python==8.3.0

passlib[bcrypt]==1.7.4
bcrypt==4.0.1

python-jose==3.3.0
python-multipart==0.0.22

pydantic==2.6.4
email-validator==2.1.0.post1
python-dotenv==1.0.1